# swagger.yaml
openapi: '3.0.3'
info:
  title: Escopo - Trabalho de Conclusão de Curso SENAI
  version: '1.0'
  description: Documentação das rotas da API do projeto Escopo.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Erro:
      type: object
      properties:
        codigo:
          type: integer
        status:
          type: string
        mensagem:
          type: string
        rota:
          type: string
        data_hora:
          type: string
    ErroCampos:
      type: object
      properties:
        codigo:
          type: integer
        status:
          type: string
        erros:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensagem:
                type: string
        rota:
          type: string
        data_hora:
          type: string
...
    UsuarioList:
      type: object
      properties:
        id:
          type: number
          example: 1
        nome:
          type: string
          example: Nathan
        email:
          type: string
          example: nathan@email.com
        foto_perfil:
          type: string
          example: https://example.com/image.jpg
...
paths:
...
  /api/v1/usuario/nome:
    patch:
      tags:
        - Usuário
      summary: Atualiza o nome do usuário autenticado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
              properties:
                nome:
                  type: string
                  example: Nathan
                  minLength: 1
                  maxLength: 100
      responses:
        '204':
          description: Nome alterado
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '400':
          description: Requisição mal formulada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: Nome inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCampos'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/usuario:
    delete:
      tags:
        - Usuário
      summary: Exclui o usuário autenticado
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Usuário excluído
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/usuarios:
    get:
      tags:
        - Usuário
      summary: Lista 5 usuários que contenham em alguma parte do email a query
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 150
          description: Filtrar usuários que contenham no email esse conteúdo
      responses:
        '200':
          description: Lista de usuários retornada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioList'
        '400':
          description: 'Parâmetro "email" não enviado, ou não atendendo aos critérios'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...